* Fix upgrade bug.

+++ system/module/upgrade/model.php
@@ -1,3 +1,4 @@
+<?php if(!defined("RUN_MODE")) die();?>
 <?php
 /**
  * The model file of upgrade module of chanzhiEPS.
@@ -152,7 +153,7 @@ class upgradeModel extends model
             case '5_3_3';
             case '5_3_4':
                 $this->fixCustomConfig();
-                $this->execSQL($this->getUpgradeFile('5.3.1'));
+                $this->execSQL($this->getUpgradeFile('5.3.4'));
             default: if(!$this->isError()) $this->loadModel('setting')->updateVersion($this->config->version);
         }
 
@@ -214,6 +215,7 @@ class upgradeModel extends model
             case '5_3_1'    : $confirmContent .= file_get_contents($this->getUpgradeFile('5.3.1'));
             case '5_3_2'    ; 
             case '5_3_3'    ; 
+            case '5_3_4'    : $confirmContent .= file_get_contents($this->getUpgradeFile('5.3.4'));
         }
         return str_replace(array('xr_', 'eps_'), $this->config->db->prefix, $confirmContent);
     }
@@ -2042,17 +2044,17 @@ class upgradeModel extends model
     public function fixCustomConfig()
     {
         $setting = $this->dao->setAutolang(false)->select('`key`, value')->from(TABLE_CONFIG)
-            ->where('oner')->eq('system')
-            ->andWere('modeul')->eq('common')
-            ->andWere('section')->eq('site')
+            ->where('owner')->eq('system')
+            ->andWhere('module')->eq('common')
+            ->andWhere('section')->eq('site')
             ->andWhere('`key`')->in('lang,requestType,defaultLang,cn2tw')
             ->fetchPairs();
 
         $this->loadModel('site')->setSystem($setting);
         $this->dao->setAutolang(false)->delete()->from(TABLE_CONFIG)
-            ->where('oner')->eq('system')
-            ->andWere('modeul')->eq('common')
-            ->andWere('section')->eq('site')
+            ->where('owner')->eq('system')
+            ->andWhere('module')->eq('common')
+            ->andWhere('section')->eq('site')
             ->andWhere('`key`')->in('lang,requestType,defaultLang,cn2tw')
             ->exec();
     }
